gif handler with variant cycle count on header

+++ gopl01/server/lissajous.go
@@ -18,7 +18,7 @@ const (
 	blackIndex = 1 // next color in palette
 )
 
-func lissajous(out io.Writer, cycles float64) {
+func lissajous(out io.Writer, cycles int) {
 	const (
 		res     = 0.001 // angular resolution
 		size    = 100   // image canvas covers [-size..+size]
@@ -32,7 +32,7 @@ func lissajous(out io.Writer, cycles float64) {
 	for i := 0; i < nframes; i++ {
 		rect := image.Rect(0, 0, 2*size+1, 2*size+1)
 		img := image.NewPaletted(rect, palette)
-		for t := 0.0; t < cycles*2*math.Pi; t += res {
+		for t := 0.0; t < float64(cycles)*2*math.Pi; t += res {
 			x := math.Sin(t)
 			y := math.Sin(t*freq + phase)
 			img.SetColorIndex(size+int(x*size+0.5), size+int(y*size+0.5),
